Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==2.2.2
requests==2.28.1
pylint==2.15.9
gunicorn==20.1.0
pymongo==4.3.3
Flask-PyMongo==2.3.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
jaeger-client==4.8.0
Flask-OpenTracing==1.1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation
opentelemetry-distro
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests